# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The contract shape for every machine-readable API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "status",
          "context"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShortCreate": {
        "properties": {
          "end_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "End Seconds",
            "type": "number"
          },
          "media_id": {
            "format": "uuid",
            "title": "Media Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ShortSource",
            "default": "manual"
          },
          "start_seconds": {
            "minimum": 0.0,
            "title": "Start Seconds",
            "type": "number"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "media_id",
          "title",
          "start_seconds",
          "end_seconds"
        ],
        "title": "ShortCreate",
        "type": "object"
      },
      "ShortResponse": {
        "properties": {
          "average_stars": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Stars"
          },
          "comment_count": {
            "title": "Comment Count",
            "type": "integer"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "end_seconds": {
            "title": "End Seconds",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_id": {
            "format": "uuid",
            "title": "Media Id",
            "type": "string"
          },
          "media_title": {
            "title": "Media Title",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ShortSource"
          },
          "start_seconds": {
            "title": "Start Seconds",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "media_id",
          "media_title",
          "title",
          "start_seconds",
          "end_seconds",
          "duration_seconds",
          "source",
          "average_stars",
          "comment_count"
        ],
        "title": "ShortResponse",
        "type": "object"
      },
      "ShortSource": {
        "enum": [
          "marker",
          "manual"
        ],
        "title": "ShortSource",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Pornarr",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/shorts": {
      "post": {
        "operationId": "admin_create_short",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Short",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/shorts/{short_id}": {
      "delete": {
        "operationId": "admin_delete_short",
        "parameters": [
          {
            "in": "path",
            "name": "short_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Short Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Short",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/shorts": {
      "get": {
        "operationId": "shorts_list_shorts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShortResponse"
                  },
                  "title": "Response Shorts List Shorts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Shorts",
        "tags": [
          "shorts"
        ]
      }
    },
    "/api/shorts/{short_id}": {
      "get": {
        "operationId": "shorts_read_short",
        "parameters": [
          {
            "in": "path",
            "name": "short_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Short Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Short",
        "tags": [
          "shorts"
        ]
      }
    },
...
  }
}